Update allo for compression and temporary files
The original script was wasting a huge amount of disk space using mainly sam files instead of bam for ease of parsing purposes. I changed these sam files to gzip compressed sam files and located them in the temporary directory /tmp, where they will be automatically removed, instead of the current wd. This allows to be more versatile in file systems where the cwd is not suitable for temporary storage of large files.